[REFACTOR] moving serializers to interfaces package level
