+++ src/interfaces/serializers/__init__.py
@@ -1,0 +1,28 @@
+from .effector import (EffectorDetailSerializer,
+                       EffectorHighlightSerializer,
+                       EffectorSerializer)
+from .evidence import (EvidenceDetailSerializer,
+                       EvidenceSerializer)
+from .gene import (GeneDetailSerializer,
+                   GeneHighlightSerializer,
+                   GeneListSerializer,
+                   GeneSerializer)
+from .operon import (OperonDetailSerializer,
+                     OperonSerializer)
+from .organism import (OrganismDetailSerializer,
+                       OrganismHighlightSerializer,
+                       OrganismSerializer)
+from .pathway import (PathwayDetailSerializer,
+                      PathwaySerializer)
+from .publication import (PublicationDetailSerializer,
+                          PublicationSerializer)
+from .regulator import (RegulatorDetailSerializer,
+                        RegulatorHighlightSerializer,
+                        RegulatorSerializer)
+from .regulatory_family import (RegulatoryFamilyDetailSerializer,
+                                RegulatoryFamilySerializer)
+from .regulatory_interaction import (RegulatoryInteractionDetailSerializer,
+                                     RegulatoryInteractionSerializer)
+from .tfbs import (TFBSDetailSerializer,
+                   TFBSHighlightSerializer,
+                   TFBSSerializer)
